+++ src/main/java/medicaldoctor/backend/data/ChangePasswordResult.java
@@ -1,4 +1,4 @@
-package medicaldoctor.backend.results;
+package medicaldoctor.backend.data;
 
 public enum ChangePasswordResult {
     PASSWORDS_NOT_MATCHING, PASSWORD_NOT_CHANGED, SUCCESS
