+++ src/main/java/medicaldoctor/backend/data/LoginResult.java
@@ -1,4 +1,4 @@
-package medicaldoctor.backend.results;
+package medicaldoctor.backend.data;
 
 public enum LoginResult {
     WRONG_USERNAME, WRONG_PASSWORD, SUCCESS;
