+++ src/test/java/medicaldoctor/tests/backend/WhenLoggingOn.java
@@ -1,6 +1,6 @@
 package medicaldoctor.tests.backend;
 
-import medicaldoctor.backend.results.LoginResult;
+import medicaldoctor.backend.data.LoginResult;
 import medicaldoctor.backend.LoginService;
 import medicaldoctor.core.AppSession;
 import medicaldoctor.entities.User;
